requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.11
idna==3.3
lxml==4.7.1
numpy==1.22.2
pandas==1.4.0
pep8==1.7.1
Pillow==9.0.1
pycodestyle==2.8.0
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2021.3
requests==2.27.1
six==1.16.0
urllib3==1.26.8
